// src/client/types.gen.ts
export type CollectiveCreateData = {
	body: {
		/**
		 * Name of the collective
		 */
		name: string
		/**
		 * Optional emoji
		 */
		emoji?: string | null
	}
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/collectives'
}
// ...
export type CollectiveUpdateData = {
	body?: {
		/**
		 * Optional emoji
		 */
		emoji?: string | null
	}
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
		/**
		 * ID of the collective
		 */
		id: number
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/collectives/{id}'
}